update api provider
