+++ fastchat/serve/gradio_web_server.py
@@ -18,6 +18,12 @@ from fastchat.conversation import SeparatorStyle
 from fastchat.constants import LOGDIR, WORKER_API_TIMEOUT, ErrorCode
 from fastchat.model.model_adapter import get_conversation_template
 from fastchat.model.model_registry import model_info
+from fastchat.serve.api_provider import (
+    anthropic_api_stream_iter,
+    bard_api_stream_iter,
+    openai_api_stream_iter,
+    palm_api_stream_iter,
+)
 from fastchat.serve.gradio_patch import Chatbot as grChatbot
 from fastchat.serve.gradio_css import code_highlight_css
 from fastchat.utils import (
@@ -203,136 +209,6 @@ def post_process_code(code):
     return code
 
 
-def openai_api_stream_iter(model_name, messages, temperature, top_p, max_new_tokens):
-    import openai
-
-    # Make requests
-    gen_params = {
-        "model": model_name,
-        "prompt": messages,
-        "temperature": temperature,
-        "top_p": top_p,
-    }
-    logger.info(f"==== request ====\n{gen_params}")
-
-    res = openai.ChatCompletion.create(
-        model=model_name, messages=messages, temperature=temperature, stream=True
-    )
-    text = ""
-    for chunk in res:
-        text += chunk["choices"][0]["delta"].get("content", "")
-        data = {
-            "text": text,
-            "error_code": 0,
-        }
-        yield data
-
-
-def anthropic_api_stream_iter(model_name, prompt, temperature, top_p, max_new_tokens):
-    import anthropic
-
-    c = anthropic.Client(os.environ["ANTHROPIC_API_KEY"])
-
-    # Make requests
-    gen_params = {
-        "model": model_name,
-        "prompt": prompt,
-        "temperature": temperature,
-        "top_p": top_p,
-    }
-    logger.info(f"==== request ====\n{gen_params}")
-
-    res = c.completion_stream(
-        prompt=prompt,
-        stop_sequences=[anthropic.HUMAN_PROMPT],
-        max_tokens_to_sample=max_new_tokens,
-        temperature=temperature,
-        top_p=top_p,
-        model=model_name,
-        stream=True,
-    )
-    for chunk in res:
-        data = {
-            "text": chunk["completion"],
-            "error_code": 0,
-        }
-        yield data
-
-
-def bard_api_stream_iter(state):
-    # TODO: we will use the official PaLM 2 API sooner or later,
-    # and we will update this function accordingly. So here we just hard code the
-    # Bard worker address. It is going to be deprecated anyway.
-
-    # Make requests
-    gen_params = {
-        "model": "bard",
-        "prompt": state.messages,
-    }
-    logger.info(f"==== request ====\n{gen_params}")
-
-    response = requests.post(
-        "http://localhost:18900/chat",
-        json={
-            "content": state.messages[-2][-1],
-            "state": state.session_state,
-        },
-        stream=False,
-        timeout=WORKER_API_TIMEOUT,
-    )
-    resp_json = response.json()
-    state.session_state = resp_json["state"]
-    content = resp_json["content"]
-    # The Bard Web API does not support streaming yet. Here we have to simulate
-    # the streaming behavior by adding some time.sleep().
-    pos = 0
-    while pos < len(content):
-        # This is a fancy way to simulate token generation latency combined
-        # with a Poisson process.
-        pos += random.randint(1, 5)
-        time.sleep(random.expovariate(20))
-        data = {
-            "text": content[:pos],
-            "error_code": 0,
-        }
-        yield data
-
-
-def init_palm_chat(model_name):
-    import vertexai
-    from vertexai.preview.language_models import ChatModel
-
-    project_id = os.environ["GCP_PROJECT_ID"]
-    location = "us-central1"
-    vertexai.init(project=project_id, location=location)
-
-    chat_model = ChatModel.from_pretrained(model_name)
-    chat = chat_model.start_chat(examples=[])
-    return chat
-
-
-def palm_api_stream_iter(chat, messages, temperature, top_p, max_new_tokens):
-    parameters = {
-        "temperature": temperature,
-        "max_output_tokens": max_new_tokens,
-        "top_p": top_p,
-    }
-    response = chat.send_message(messages, **parameters)
-    content = response.text
-
-    pos = 0
-    while pos < len(content):
-        # This is a fancy way to simulate token generation latency combined
-        # with a Poisson process.
-        pos += random.randint(1, 5)
-        time.sleep(random.expovariate(20))
-        data = {
-            "text": content[:pos],
-            "error_code": 0,
-        }
-        yield data
-
-
 def model_worker_stream_iter(
     conv, model_name, worker_addr, prompt, temperature, top_p, max_new_tokens
 ):
@@ -408,6 +284,7 @@ def http_bot(
             model_name, prompt, temperature, top_p, max_new_tokens
         )
     elif model_name == "bard":
+        #stream_iter = bard_api_stream_iter(state)
         stream_iter = palm_api_stream_iter(
             state.chat, state.messages[-2][1], temperature, top_p, max_new_tokens
         )
