+++ fastchat/serve/huggingface_api.py
@@ -1,4 +1,6 @@
 """
+Use FastChat with Hugging Face generation APIs.
+
 Usage:
 python3 -m fastchat.serve.huggingface_api --model lmsys/fastchat-t5-3b-v1.0
 python3 -m fastchat.serve.huggingface_api --model ~/model_weights/vicuna-7b/
