+++ fastchat/serve/test_message.py
@@ -1,3 +1,4 @@
+"""Send a test message."""
 import argparse
 import json
